client/requirements.txt
click==6.7
ruamel.yaml==0.14.9
requests==2.18.4